+++ packages/dumb/src/components/stats/time-in-range/time-in-range-chart/time-in-range-chart.hook.test.tsx
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2022-2025, Diabeloop
+ * Copyright (c) 2022-2026, Diabeloop
  *
  * All rights reserved.
  *
@@ -32,13 +32,14 @@ import { type CBGPercentageData, CBGStatType, StatLevel } from '../../../../mode
 import { type TimeInRangeData } from 'tidepool-viz/src/types/utils/data'
 import { DatumType, Unit } from 'medical-domain'
 
-describe('CBGPercentageBarChart hook', () => {
+describe('Time in range chart hook', () => {
   const veryHighStatValue = 100
   const highStatValue = 200
   const targetStatValue = 150
   const lowStatValue = 250
   const veryLowStatValue = 50
   const total = 1000
+
   const createCBGTimeData = (id: StatLevel, title: string, value: number): CBGPercentageData => {
     return { id, title, value }
   }
@@ -83,6 +84,7 @@ describe('CBGPercentageBarChart hook', () => {
   it('should return correct cbgStatsProps', () => {
     const props = { ...defaultProps }
     const { result } = renderHook(() => useTimeInRangeChartHook(props))
+
     expect(result.current.cbgStatsProps).toEqual({
       veryHighStat: {
         type: CBGStatType.TimeInRange,
@@ -131,22 +133,9 @@ describe('CBGPercentageBarChart hook', () => {
 
     const { result: thirdHook } = renderHook(() => useTimeInRangeChartHook({
       ...defaultProps,
-      type: CBGStatType.ReadingsInRange
-    }))
-    expect(thirdHook.current.title).toEqual('Avg. Daily Readings In Range')
-
-    const { result: fourthHook } = renderHook(() => useTimeInRangeChartHook({
-      ...defaultProps,
-      days: 0,
-      type: CBGStatType.ReadingsInRange
-    }))
-    expect(fourthHook.current.title).toEqual('Readings In Range')
-
-    const { result: fifthHook } = renderHook(() => useTimeInRangeChartHook({
-      ...defaultProps,
       bgType: DatumType.Smbg
     }))
-    expect(fifthHook.current.title).toEqual('Time In Range')
+    expect(thirdHook.current.title).toEqual('Time In Range')
   })
 
   it('should compute the right annotations', () => {
@@ -158,16 +147,9 @@ describe('CBGPercentageBarChart hook', () => {
 
     const { result: thirdHook } = renderHook(() => useTimeInRangeChartHook({
       ...defaultProps,
-      days: 0,
-      type: CBGStatType.ReadingsInRange
-    }))
-    expect(thirdHook.current.annotations).toEqual(['**Readings In Range:** Number of BGM readings.'])
-
-    const { result: fourthHook } = renderHook(() => useTimeInRangeChartHook({
-      ...defaultProps,
       bgType: DatumType.Smbg
     }))
-    expect(fourthHook.current.annotations).toEqual(['**Time In Range:** Time spent in range, based on CGM readings.', '**How we calculate this:**\n\n**(%)** is the number of readings in range divided by all readings for this time period.\n\n**(time)** is 24 hours multiplied by % in range.', 'Derived from _**1000**_ BGM readings.'])
+    expect(thirdHook.current.annotations).toEqual(['**Time In Range:** Time spent in range, based on CGM readings.', '**How we calculate this:**\n\n**(%)** is the number of readings in range divided by all readings for this time period.\n\n**(time)** is 24 hours multiplied by % in range.', 'Derived from _**1000**_ BGM readings.'])
   })
 
   it('onMouseOver and OnMouseLeave should return correct values', async () => {